Updated model files with required fields
